refactor: use WeakMaps
